Revert "halve entry size"
This reverts commit 771aec587f8108fa841248a630330236d4c061cb.

bench: 1947240

+++ src/engine/search/history.h
@@ -16,7 +16,7 @@ namespace episteme::hist {
     }
 
     struct Entry {
-        int16_t value = 0;
+        int32_t value = 0;
 
         inline void update(int16_t bonus, int32_t max) {
             value += bonus - value * std::abs(bonus) / max;
